feat: Add missing commit and retry statement mechanism.

+++ internal/repository/session_repository.go
@@ -2,6 +2,7 @@ package repository
 
 import (
 	"database/sql"
+	"github.com/cockroachdb/cockroach-go/v2/crdb"
 	"social-media/internal/entity"
 )
 
@@ -37,10 +38,14 @@ func DeserializeSessionRows(rows *sql.Rows) []*entity.Session {
 }
 
 func (sessionRepository *SessionRepository) FindOneById(begin *sql.Tx, id string) *entity.Session {
-	rows, queryErr := begin.Query(
-		"SELECT id, user_id, access_token, refresh_token, access_token_expired_at, refresh_token_expired_at, created_at, updated_at, deleted_at FROM \"session\" WHERE id=$1 LIMIT 1;",
-		id,
-	)
+	var rows *sql.Rows
+	queryErr := crdb.Execute(func() (err error) {
+		rows, err = begin.Query(
+			"SELECT id, user_id, access_token, refresh_token, access_token_expired_at, refresh_token_expired_at, created_at, updated_at, deleted_at FROM \"session\" WHERE id=$1 LIMIT 1;",
+			id,
+		)
+		return err
+	})
 	if queryErr != nil {
 		panic(queryErr)
 	}
@@ -54,10 +59,14 @@ func (sessionRepository *SessionRepository) FindOneById(begin *sql.Tx, id string
 }
 
 func (sessionRepository *SessionRepository) FindOneByUserId(begin *sql.Tx, userId string) *entity.Session {
-	rows, queryErr := begin.Query(
-		"SELECT id, user_id, access_token, refresh_token, access_token_expired_at, refresh_token_expired_at, created_at, updated_at, deleted_at FROM \"session\" WHERE user_id=$1 LIMIT 1;",
-		userId,
-	)
+	var rows *sql.Rows
+	queryErr := crdb.Execute(func() (err error) {
+		rows, err = begin.Query(
+			"SELECT id, user_id, access_token, refresh_token, access_token_expired_at, refresh_token_expired_at, created_at, updated_at, deleted_at FROM \"session\" WHERE user_id=$1 LIMIT 1;",
+			userId,
+		)
+		return err
+	})
 	if queryErr != nil {
 		panic(queryErr)
 	}
@@ -71,18 +80,21 @@ func (sessionRepository *SessionRepository) FindOneByUserId(begin *sql.Tx, userI
 }
 
 func (sessionRepository *SessionRepository) CreateOne(begin *sql.Tx, toCreateSession *entity.Session) *entity.Session {
-	_, queryErr := begin.Query(
-		"INSERT INTO \"session\" (id, user_id, access_token, refresh_token, access_token_expired_at, refresh_token_expired_at, created_at, updated_at, deleted_at) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9);",
-		toCreateSession.Id,
-		toCreateSession.UserId,
-		toCreateSession.AccessToken,
-		toCreateSession.RefreshToken,
-		toCreateSession.AccessTokenExpiredAt,
-		toCreateSession.RefreshTokenExpiredAt,
-		toCreateSession.CreatedAt,
-		toCreateSession.UpdatedAt,
-		toCreateSession.DeletedAt,
-	)
+	queryErr := crdb.Execute(func() (err error) {
+		_, err = begin.Query(
+			"INSERT INTO \"session\" (id, user_id, access_token, refresh_token, access_token_expired_at, refresh_token_expired_at, created_at, updated_at, deleted_at) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9);",
+			toCreateSession.Id,
+			toCreateSession.UserId,
+			toCreateSession.AccessToken,
+			toCreateSession.RefreshToken,
+			toCreateSession.AccessTokenExpiredAt,
+			toCreateSession.RefreshTokenExpiredAt,
+			toCreateSession.CreatedAt,
+			toCreateSession.UpdatedAt,
+			toCreateSession.DeletedAt,
+		)
+		return err
+	})
 	if queryErr != nil {
 		panic(queryErr)
 	}
@@ -91,70 +103,45 @@ func (sessionRepository *SessionRepository) CreateOne(begin *sql.Tx, toCreateSes
 }
 
 func (sessionRepository *SessionRepository) PatchOneById(begin *sql.Tx, id string, toPatchSession *entity.Session) *entity.Session {
-	foundRows, foundRowsErr := begin.Query(
-		"SELECT id, user_id, access_token, refresh_token, access_token_expired_at, refresh_token_expired_at, created_at, updated_at, deleted_at FROM \"session\" WHERE id=$1 LIMIT 1;",
-		id,
-	)
-	if foundRowsErr != nil {
-		panic(foundRowsErr)
-	}
-
-	foundSessions := DeserializeSessionRows(foundRows)
-	if len(foundSessions) == 0 {
-		return nil
-	}
-
-	foundSession := foundSessions[0]
-	foundSession.Id = toPatchSession.Id
-	foundSession.UserId = toPatchSession.UserId
-	foundSession.AccessToken = toPatchSession.AccessToken
-	foundSession.RefreshToken = toPatchSession.RefreshToken
-	foundSession.AccessTokenExpiredAt = toPatchSession.AccessTokenExpiredAt
-	foundSession.RefreshTokenExpiredAt = toPatchSession.RefreshTokenExpiredAt
-	foundSession.CreatedAt = toPatchSession.CreatedAt
-	foundSession.UpdatedAt = toPatchSession.UpdatedAt
-	foundSession.DeletedAt = toPatchSession.DeletedAt
-
-	_, queryErr := begin.Query(
-		"UPDATE \"session\" SET id=$1, user_id=$2, access_token=$3, refresh_token=$4, access_token_expired_at=$5, refresh_token_expired_at=$6, created_at=$7, updated_at=$8, deleted_at=$9 WHERE id=$10 LIMIT 1;",
-		foundSession.Id,
-		foundSession.UserId,
-		foundSession.AccessToken,
-		foundSession.RefreshToken,
-		foundSession.AccessTokenExpiredAt,
-		foundSession.RefreshTokenExpiredAt,
-		foundSession.CreatedAt,
-		foundSession.UpdatedAt,
-		foundSession.DeletedAt,
-		id,
-	)
+	queryErr := crdb.Execute(func() (err error) {
+		_, err = begin.Query(
+			"UPDATE \"session\" SET id=$1, user_id=$2, access_token=$3, refresh_token=$4, access_token_expired_at=$5, refresh_token_expired_at=$6, created_at=$7, updated_at=$8, deleted_at=$9 WHERE id=$10 LIMIT 1;",
+			toPatchSession.Id,
+			toPatchSession.UserId,
+			toPatchSession.AccessToken,
+			toPatchSession.RefreshToken,
+			toPatchSession.AccessTokenExpiredAt,
+			toPatchSession.RefreshTokenExpiredAt,
+			toPatchSession.CreatedAt,
+			toPatchSession.UpdatedAt,
+			toPatchSession.DeletedAt,
+			id,
+		)
+		return err
+	})
 	if queryErr != nil {
 		panic(queryErr)
 	}
 
-	return foundSession
+	return toPatchSession
 }
 
 func (sessionRepository *SessionRepository) DeleteOneById(begin *sql.Tx, id string) *entity.Session {
-	foundRows, foundRowsErr := begin.Query(
-		"SELECT id, user_id, access_token, refresh_token, access_token_expired_at, refresh_token_expired_at, created_at, updated_at, deleted_at FROM \"session\" WHERE id=$1 LIMIT 1;",
-		id,
-	)
-	if foundRowsErr != nil {
-		panic(foundRowsErr)
-	}
-
-	foundSessions := DeserializeSessionRows(foundRows)
-	if len(foundSessions) == 0 {
-		return nil
-	}
-
-	_, queryErr := begin.Query(
-		"DELETE FROM \"session\" WHERE id=$1",
-		id,
-	)
+	var rows *sql.Rows
+	queryErr := crdb.Execute(func() (err error) {
+		rows, err = begin.Query(
+			"DELETE FROM \"session\" WHERE id=$1 LIMIT 1 RETURNING id, user_id, access_token, refresh_token, access_token_expired_at, refresh_token_expired_at, created_at, updated_at, deleted_at;",
+			id,
+		)
+		return err
+	})
 	if queryErr != nil {
 		panic(queryErr)
+	}
+
+	foundSessions := DeserializeSessionRows(rows)
+	if len(foundSessions) == 0 {
+		return nil
 	}
 
 	return foundSessions[0]
